Don't use static IP, add alias to https-proxy
Using a static IP address caused failure on system reboot: "Address already in use." This is because Docker apparently assigned the desired IP to another container before attempting to allocate an IP for the https-proxy.

Instead, we are now using an alias, https-proxy which we can resolve in the Drupal configuration.